+++ src/main/java/com/example/clearthefridge/domain/recipe/entity/RecipeIngredient.java
@@ -1,0 +1,28 @@
+package com.example.clearthefridge.domain.recipe.entity;
+
+import com.example.clearthefridge.domain.ingredient.entity.Ingredient;
+import com.example.clearthefridge.domain.recipe.entity.Recipe;
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Table(name = "recipe_ingredients")
+@Getter
+@AllArgsConstructor
+@NoArgsConstructor
+public class RecipeIngredient {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @ManyToOne
+    private Recipe recipe;
+
+    @ManyToOne
+    private Ingredient ingredient;
+
+    private String amount;
+}
