+++ src/main/java/com/example/clearthefridge/domain/recipe/repository/exam.java
@@ -1,0 +1,4 @@
+package com.example.clearthefridge.domain.recipe.repository;
+
+public class exam {
+}
