+++ src/main/java/com/example/clearthefridge/global/config/WebConfig.java
@@ -1,0 +1,4 @@
+package com.example.clearthefridge.global.config;
+
+public class WebConfig {
+}
