--- src/main/java/com/example/clearthefridge/ingredient/entity/Ingredient.java
@@ -1,26 +1,0 @@
-package com.example.clearthefridge.ingredient.entity;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Table(name = "ingredients")
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-public class Ingredient {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    //재료 이름
-    private String name;
-
-    //소비 권장 기한
-    private Integer lifeDays;
-}
