--- src/main/java/com/example/clearthefridge/ingredient/service/IngredientService.java
@@ -1,7 +1,0 @@
-package com.example.clearthefridge.ingredient.service;
-
-
-
-public interface IngredientService {
-
-}
