--- src/main/java/com/example/clearthefridge/recipeIngredient/entity/RecipeIngredient.java
@@ -1,28 +1,0 @@
-package com.example.clearthefridge.recipeIngredient.entity;
-
-import com.example.clearthefridge.ingredient.entity.Ingredient;
-import com.example.clearthefridge.recipe.entity.Recipe;
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Table(name = "recipe_ingredients")
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor
-public class RecipeIngredient {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne
-    private Recipe recipe;
-
-    @ManyToOne
-    private Ingredient ingredient;
-
-    private String amount;
-}
